requirements.txt
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
Flask==0.10.1
github3.py==0.9.4
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.29
MarkupSafe==0.23
pytz==2015.7
redis==2.10.5
requests==2.8.1
uritemplate.py==0.3.0
Werkzeug==0.11.2
